services/agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-groq
langchain_tavily
langchain-community
langgraph
python-dotenv